+++ examples/play_track.py
@@ -3,8 +3,8 @@
 """
 This is an example of playing music from Spotify using pyspotify.
 
-The example use the AlsaDriver, and will thus only work on systems with an ALSA
-sound subsystem, which means most Linux systems.
+The example use the :class:`spotify.sink.AlsaSink`, and will thus only work on
+systems with an ALSA sound subsystem, which means most Linux systems.
 
 You can either run this file directly without arguments to play a default
 track::
@@ -22,7 +22,7 @@ import sys
 import threading
 
 import spotify
-import spotify.alsa
+import spotify.sink
 
 if sys.argv[1:]:
     track_uri = sys.argv[1]
@@ -36,8 +36,8 @@ session = spotify.Session()
 loop = spotify.EventLoop(session)
 loop.start()
 
-# Connect an audio driver
-audio = spotify.alsa.AlsaDriver(session)
+# Connect an audio sink
+audio = spotify.sink.AlsaSink(session)
 
 # EVents for coordination
 logged_in = threading.Event()