+++ examples/shell.py
@@ -20,9 +20,9 @@ import threading
 import spotify
 
 try:
-    from spotify.alsa import AlsaDriver
+    from spotify.sink import AlsaSink
 except ImportError:
-    AlsaDriver = None
+    AlsaSink = None
 
 
 class Commander(cmd.Cmd):
@@ -45,8 +45,11 @@ class Commander(cmd.Cmd):
         self.session.on(
             spotify.SessionEvent.END_OF_TRACK, self.on_end_of_track)
 
-        if AlsaDriver is not None:
-            self.audio_driver = AlsaDriver(self.session)
+        if AlsaSink is not None:
+            self.audio_driver = AlsaSink(self.session)
+        else:
+            self.logger.warning(
+                'No audio sink found; audio playback unavailable.')
 
         self.event_loop = spotify.EventLoop(self.session)
         self.event_loop.start()